+++ example_client.py
@@ -11,14 +11,13 @@ logging.basicConfig(
 
 async def test(port=3755):
     print("connecting to broker")
-    client = await RpcClient.connect(
-        host="localhost",
-        port=port,
+    client = await RpcClient.connect(host="localhost", port=port)
+    print("connected OK")
+    await client.login(
         password="test",
         user="test",
         login_type=RpcClient.LoginType.PLAIN,
     )
-    print("connected OK")
     print("calling shv method 'echo'")
     await client.call_shv_method(".broker/app", "echo", 42)
     resp = await client.read_rpc_message()